+++ src/main/java/com/backend/controller/MovieController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.backend.model.TMDB.MovieDto;
+import com.backend.dto.TMDB.MovieDto;
 import com.backend.service.TMDBService;
 
 @RestController
