+++ src/main/java/com/backend/dto/AuthRequest.java
@@ -1,4 +1,4 @@
-package com.backend.model;
+package com.backend.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,8 +7,7 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class GenreDTO {
-    private Long id;
-    private String name;
-    private String description;
+public class AuthRequest {
+	private String email;
+    private String password;
 }
