+++ src/main/java/com/backend/dto/AuthResponse.java
@@ -1,4 +1,4 @@
-package com.backend.model;
+package com.backend.dto;
 
 import lombok.*;
