+++ src/main/java/com/backend/dto/ChangePassword.java
@@ -1,4 +1,4 @@
-package com.backend.model;
+package com.backend.dto;
 
 public record ChangePassword(String password, String repeatPassword) {
