+++ src/main/java/com/backend/dto/TMDB/TMDbSearchTVResponse.java
@@ -1,0 +1,10 @@
+package com.backend.dto.TMDB;
+
+import java.util.List;
+
+import lombok.Data;
+
+@Data
+public class TMDbSearchTVResponse {
+    private List<TVDto> results;
+}
